src/Frontend/requirements.txt
transformers
pip
jax
flax
flask
paho_mqtt